+++ routes/like.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const request = require('../database/likes.js');
 const router = express.Router();
+const auth = require('../middleware/authentification')
 
 // Route to get the number of likes for a course
 router.get('/:id_course/:id_utilisateur', async (req, res) => {
@@ -8,7 +9,6 @@ router.get('/:id_course/:id_utilisateur', async (req, res) => {
         const id_course = req.params.id_course
         const id_utilisateur = req.params.id_utilisateur
 
-        console.log(id_course, id_utilisateur)
         const data = await request.getLike(id_course, id_utilisateur)
         
         console.log(data)
@@ -23,12 +23,14 @@ router.get('/:id_course/:id_utilisateur', async (req, res) => {
 })
 
 // Route to add a like to a course
-router.post('/:id_course/:id_utilisateur', async (req, res) => {
+router.post('/:id_course/:id_utilisateur', auth, async (req, res) => {
     try {
         const id_course = req.params.id_course
         const id_utilisateur = req.params.id_utilisateur
 
-        console.log(id_course, id_utilisateur)
+        // Authentification de l'Utilisateur
+        if (req.user.id != id_utilisateur) return res.status(401).json({ message: "Unauthorized."})
+
         const data = await request.addLike(id_course, id_utilisateur)
         
         res.status(200).json({ message: "success" })
@@ -38,12 +40,14 @@ router.post('/:id_course/:id_utilisateur', async (req, res) => {
 })
 
 // Route to delete a like from a course
-router.delete('/:id_course/:id_utilisateur', async (req, res) => {
+router.delete('/:id_course/:id_utilisateur', auth, async (req, res) => {
     try {
         const id_course = req.params.id_course
         const id_utilisateur = req.params.id_utilisateur
 
-        console.log(id_course, id_utilisateur)
+        // Authentification de l'Utilisateur
+        if (req.user.id != id_utilisateur) return res.status(401).json({ message: "Unauthorized."})
+
         const data = await request.deleteLike(id_course, id_utilisateur)
         
         res.status(200).json({ message: "success" })
